+++ sbin/configure-environment.bash
@@ -1,4 +1,5 @@
 #!/bin/bash
+MYSCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 
-LIRETOOLS=/cygdrive/f/Projekte/liretools/
+LIRETOOLS=${MYSCRIPTPATH}/../
 W_LIREOOLS="F:\\Projekte\\liretools\\"