+++ src/main/java/de/mytb/liretools/MatchingHit.java
@@ -1,14 +1,17 @@
 package de.mytb.liretools;
 
+import lombok.Getter;
+
 import java.text.DecimalFormat;
 
+@Getter
 public class MatchingHit {
     private static final DecimalFormat DF2 = new DecimalFormat("#.##");
 
-    private Class extractor;
-    private String fileName;
-    private Double score;
-    private Double relScore;
+    private final Class extractor;
+    private final String fileName;
+    private final Double score;
+    private final Double relScore;
 
     public MatchingHit(Class extractor, String fileName, Double score, Double relScore) {
         this.extractor = extractor;
@@ -16,23 +19,6 @@ public class MatchingHit {
         this.score = score;
         this.relScore = relScore;
     }
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public Double getScore() {
-        return score;
-    }
-
-    public Double getRelScore() {
-        return relScore;
-    }
-
-    public Class getExtractor() {
-        return extractor;
-    }
-
     public String toString() {
         return "{" + getExtractor().getSimpleName() + ": " + DF2.format(getScore()) + " / " + DF2.format(getRelScore()) + "}";
     }