+++ src/main/java/de/mytb/liretools/SearchRequest.java
@@ -1,0 +1,11 @@
+package de.mytb.liretools;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SearchRequest<R> {
+    private final SearchRequestType type;
+    private final R request;
+}