+++ src/main/java/de/mytb/liretools/StdOutWriter.java
@@ -1,0 +1,10 @@
+package de.mytb.liretools;
+
+public class StdOutWriter implements Writer {
+    @Override
+    synchronized public void println(String text) {
+        System.out.println(text);
+    }
+
+
+}