+++ docs/CHANGELOG.md
@@ -2,6 +2,30 @@
  
 # Versions
  
+## 1.1.0
+- improved sbin-scripts
+- refactored code
+- added similarity-indexer
+
+### new features
+- similarity-indexer: added similarity-searcher to get a list of similar documents in index
+ 
+### improvements
+- scripts: improved sbin-scripts
+- indexer: refactored code
+- searcher: refactored code
+
+### bug fixes
+- none
+ 
+### breaking changes
+- none
+
+### known issues
+- none
+
+
+
 ## 1.0.0
 - initial version